web: separate fetching of backbone collection into higher order component
