+++ web/js/items/react-wrapper.js
@@ -1,11 +1,10 @@
 'use strict';
 
 var React = require('react');
-var ReactDOM = require('react-dom');
 var ItemModel = require('./models/itemModel');
 var ItemsCollection = require('./models/itemsCollection');
 var CostsTable = require('./CostsTable.jsx');
-var Loader = require('../components/Loader.jsx');
+var CollectionLoader = require('../components/CollectionLoader.jsx');
 require('react.backbone');
 
 var ItemComponent = React.createBackboneClass({
@@ -118,40 +117,11 @@ var ItemsListComponent = React.createBackboneClass({
   }
 });
 
-// TODO
-// This should probably become react mixin or something like that.
-// Also, look for beter ways of data initialization.
-module.exports = React.createClass({
-  getInitialState: function () {
-    return {
-      error: null,
-      loading: true,
-      collection: ItemsCollection.singleton()
-    };
-  },
-
-  componentDidMount: function () {
-    this.state.collection.fetch({
-      reset: true,
-      success: function () {
-        this.setState({loading: false});
-      }.bind(this),
-      error: function (collection, xhr) {
-        var err = new Error(xhr.responseText);
-        err.reason = xhr.responseJSON;
-        this.setState({
-          loading: false,
-          error: err
-        });
-      }.bind(this)
-    });
-  },
-
-  render: function () {
-    return (
-      <Loader loading={this.state.loading} error={this.state.error}>
-        <ItemsListComponent collection={this.state.collection} />
-      </Loader>
-    );
-  }
-});
+module.exports = function () {
+  return (
+    <CollectionLoader
+      collection={ItemsCollection.singleton()}
+      component={ItemsListComponent}
+    />
+  );
+};
